feat: add visual decimal precision to 5-hour usage percentage
The API now returns decimal values (e.g., 5.3%) instead of integers. Display
the decimal portion using vertical Unicode block characters (▁▂▃▄▅▆▇█) after
the percentage sign to show fine-grained progress towards the next whole
percentage point.

Format changes from "5h:5% 7d:12%" to "5h:5%▃7d:12%" where the block character
naturally acts as a separator and visual progress indicator.

+++ src/api_usage.rs
@@ -119,8 +119,8 @@ fn fetch_usage_internal() -> Result<ApiUsageData> {
         .and_then(|s| s.parse::<DateTime<Utc>>().ok());
 
     Ok(ApiUsageData {
-        five_hour_percent: api_response.five_hour.utilization as u32,
+        five_hour_percent: api_response.five_hour.utilization,
         five_hour_resets_at,
-        seven_day_percent: api_response.seven_day.utilization as u32,
+        seven_day_percent: api_response.seven_day.utilization,
     })
 }
