+++ src/types.rs
@@ -170,7 +170,7 @@ pub struct ContextInfo {
 /// API usage data from claude.ai
 #[derive(Debug, Clone)]
 pub struct ApiUsageData {
-    pub five_hour_percent: u32,
+    pub five_hour_percent: f64,
     pub five_hour_resets_at: Option<DateTime<Utc>>,
-    pub seven_day_percent: u32,
+    pub seven_day_percent: f64,
 }
